cclient/http: Separated sockets
