+++ network/BUILD
@@ -1,0 +1,7 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "sockets",
+    hdrs = ["sockets.h"],
+    srcs = ["sockets.c"],
+)
